#233 Refactor: WineWishlist 리패키징
