+++ src/main/java/com/drinkeg/drinkeg/converter/WineWishlistConverter.java
@@ -2,8 +2,8 @@ package com.drinkeg.drinkeg.converter;
 
 import com.drinkeg.drinkeg.domain.Member;
 import com.drinkeg.drinkeg.wine.domain.Wine;
-import com.drinkeg.drinkeg.domain.WineWishlist;
-import com.drinkeg.drinkeg.dto.WineWishlistDTO.response.WineWishlistResponseDTO;
+import com.drinkeg.drinkeg.wineWishlist.domain.WineWishlist;
+import com.drinkeg.drinkeg.wineWishlist.dto.response.WineWishlistResponseDTO;
 import com.drinkeg.drinkeg.wine.dto.response.SearchWineResponseDTO;
 
 public class WineWishlistConverter {
