+++ src/main/java/com/drinkeg/drinkeg/wineWishlist/repository/WineWishlistRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.drinkeg.drinkeg.repository;
+package com.drinkeg.drinkeg.wineWishlist.repository;
 
 public interface WineWishlistRepositoryCustom {
