+++ src/main/java/com/drinkeg/drinkeg/wineWishlist/service/WineWishlistServiceImpl.java
@@ -1,16 +1,16 @@
-package com.drinkeg.drinkeg.service.wineWishlistService;
+package com.drinkeg.drinkeg.wineWishlist.service;
 
 import com.drinkeg.drinkeg.apipayLoad.code.status.ErrorStatus;
 import com.drinkeg.drinkeg.converter.WineWishlistConverter;
 import com.drinkeg.drinkeg.domain.Member;
 import com.drinkeg.drinkeg.wine.domain.Wine;
-import com.drinkeg.drinkeg.domain.WineWishlist;
-import com.drinkeg.drinkeg.dto.WineWishlistDTO.request.WineWishlistRequestDTO;
-import com.drinkeg.drinkeg.dto.WineWishlistDTO.response.WineWishlistResponseDTO;
+import com.drinkeg.drinkeg.wineWishlist.domain.WineWishlist;
+import com.drinkeg.drinkeg.wineWishlist.dto.request.WineWishlistRequestDTO;
+import com.drinkeg.drinkeg.wineWishlist.dto.response.WineWishlistResponseDTO;
 import com.drinkeg.drinkeg.exception.GeneralException;
 import com.drinkeg.drinkeg.repository.MemberRepository;
 import com.drinkeg.drinkeg.wine.repository.WineRepository;
-import com.drinkeg.drinkeg.repository.WineWishlistRepository;
+import com.drinkeg.drinkeg.wineWishlist.repository.WineWishlistRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
